implement new way of price calculation

+++ PrintMe.Application/Models/SizeDto.cs
@@ -2,4 +2,4 @@ using PrintMe.Application.Enums;
 
 namespace PrintMe.Application.Model;
 
-public record SizeDto(int Id, string Name, string Description, long Multiplier);
+public record SizeDto(int Id, string Name, string Description, double Multiplier);
